refactor: extract HistoryItem to a shared AgentInstanceHistoryItem client command type
Promotes HistoryItem from a type nested inside
UpdateAgentInstanceCommandStep1 to a top-level command type, renamed
to AgentInstanceHistoryItem, so it can be shared with the create
command's builder without nesting it under an unrelated interface.
Pure move: no behavior change.
