+++ clients/java/src/main/java/io/camunda/client/api/command/UpdateAgentInstanceCommandStep1.java
@@ -16,8 +16,6 @@
 package io.camunda.client.api.command;
 
 import io.camunda.client.api.response.UpdateAgentInstanceResponse;
-import io.camunda.client.api.search.enums.AgentInstanceHistoryRole;
-import java.time.OffsetDateTime;
 import java.util.List;
 
 /**
@@ -135,206 +133,6 @@ public interface UpdateAgentInstanceCommandStep1 {
      *     null.
      * @return this builder for method chaining
      */
-    UpdateAgentInstanceCommandStep2 history(List<HistoryItem> history);
-  }
-
-  /** A single conversation history item to append as part of an update batch. */
-  final class HistoryItem {
-    private String historyItemId;
-    private int loopIteration;
-    private AgentInstanceHistoryRole role;
-    private List<AgentInstanceHistoryContent> content;
-    private OffsetDateTime producedAt;
-    private List<AgentInstanceHistoryToolCall> toolCalls;
-    private AgentInstanceHistoryMetrics metrics;
-    private List<AgentTool> tools;
-    private String model;
-    private String provider;
-    private AgentInstanceLimits limits;
-    private List<AgentInstanceHistoryContent> systemPrompt;
-
-    /**
-     * Sets the caller-assigned identifier used to detect and dedupe retries of the same item.
-     *
-     * @param historyItemId the item id. Must not be null or blank.
-     * @return this builder for method chaining
-     */
-    public HistoryItem historyItemId(final String historyItemId) {
-      this.historyItemId = historyItemId;
-      return this;
-    }
-
-    /**
-     * Sets the loop iteration this item belongs to: one pass through the agent's loop, during which
-     * the model reasons, selects tools, evaluates the result, and decides whether to continue.
-     * Grouping items by iteration makes a single pass addressable, so a slice of the conversation
-     * history can be pointed at rather than the whole conversation.
-     *
-     * @param loopIteration the loop iteration number. Must be greater than 0, and increase with
-     *     each iteration.
-     * @return this builder for method chaining
-     */
-    public HistoryItem loopIteration(final int loopIteration) {
-      this.loopIteration = loopIteration;
-      return this;
-    }
-
-    /**
-     * Sets the role of this history item in the conversation.
-     *
-     * @param role the conversation role. Must not be null.
-     * @return this builder for method chaining
-     */
-    public HistoryItem role(final AgentInstanceHistoryRole role) {
-      this.role = role;
-      return this;
-    }
-
-    /**
-     * Sets the content blocks of this history item.
-     *
-     * @param content the list of content blocks. Must not be null or empty.
-     * @return this builder for method chaining
-     */
-    public HistoryItem content(final List<AgentInstanceHistoryContent> content) {
-      this.content = content;
-      return this;
-    }
-
-    /**
-     * Sets the agent-side timestamp when this message was produced.
-     *
-     * @param producedAt the production timestamp. Must not be null.
-     * @return this builder for method chaining
-     */
-    public HistoryItem producedAt(final OffsetDateTime producedAt) {
-      this.producedAt = producedAt;
-      return this;
-    }
-
-    /**
-     * Sets the tool calls associated with this history item.
-     *
-     * @param toolCalls the list of tool calls. May be null.
-     * @return this builder for method chaining
-     */
-    public HistoryItem toolCalls(final List<AgentInstanceHistoryToolCall> toolCalls) {
-      this.toolCalls = toolCalls;
-      return this;
-    }
-
-    /**
-     * Sets per-call token and latency metrics. Present on ASSISTANT items only.
-     *
-     * @param metrics the metrics. May be null.
-     * @return this builder for method chaining
-     */
-    public HistoryItem metrics(final AgentInstanceHistoryMetrics metrics) {
-      this.metrics = metrics;
-      return this;
-    }
-
-    /**
-     * Sets the tools available to the agent as of this entry. CONFIGURATION items only.
-     *
-     * @param tools the tools. May be null.
-     * @return this builder for method chaining
-     */
-    public HistoryItem tools(final List<AgentTool> tools) {
-      this.tools = tools;
-      return this;
-    }
-
-    /**
-     * Sets the LLM model identifier as of this entry. CONFIGURATION items only.
-     *
-     * @param model the model identifier. May be null.
-     * @return this builder for method chaining
-     */
-    public HistoryItem model(final String model) {
-      this.model = model;
-      return this;
-    }
-
-    /**
-     * Sets the LLM provider as of this entry. CONFIGURATION items only.
-     *
-     * @param provider the provider identifier. May be null.
-     * @return this builder for method chaining
-     */
-    public HistoryItem provider(final String provider) {
-      this.provider = provider;
-      return this;
-    }
-
-    /**
-     * Sets the operational limits as of this entry. CONFIGURATION items only.
-     *
-     * @param limits the limits. May be null.
-     * @return this builder for method chaining
-     */
-    public HistoryItem limits(final AgentInstanceLimits limits) {
-      this.limits = limits;
-      return this;
-    }
-
-    /**
-     * Sets the system prompt, as content blocks, as of this entry. CONFIGURATION items only.
-     *
-     * @param systemPrompt the system prompt content blocks. May be null.
-     * @return this builder for method chaining
-     */
-    public HistoryItem systemPrompt(final List<AgentInstanceHistoryContent> systemPrompt) {
-      this.systemPrompt = systemPrompt;
-      return this;
-    }
-
-    public String getHistoryItemId() {
-      return historyItemId;
-    }
-
-    public int getLoopIteration() {
-      return loopIteration;
-    }
-
-    public AgentInstanceHistoryRole getRole() {
-      return role;
-    }
-
-    public List<AgentInstanceHistoryContent> getContent() {
-      return content;
-    }
-
-    public OffsetDateTime getProducedAt() {
-      return producedAt;
-    }
-
-    public List<AgentInstanceHistoryToolCall> getToolCalls() {
-      return toolCalls;
-    }
-
-    public AgentInstanceHistoryMetrics getMetrics() {
-      return metrics;
-    }
-
-    public List<AgentTool> getTools() {
-      return tools;
-    }
-
-    public String getModel() {
-      return model;
-    }
-
-    public String getProvider() {
-      return provider;
-    }
-
-    public AgentInstanceLimits getLimits() {
-      return limits;
-    }
-
-    public List<AgentInstanceHistoryContent> getSystemPrompt() {
-      return systemPrompt;
-    }
+    UpdateAgentInstanceCommandStep2 history(List<AgentInstanceHistoryItem> history);
   }
 }
