+++ clients/java/src/main/java/io/camunda/client/impl/command/AgentInstanceHistoryMapper.java
@@ -19,6 +19,7 @@ import io.camunda.client.api.command.AgentInstanceHistoryContent;
 import io.camunda.client.api.command.AgentInstanceHistoryContent.DocumentContent;
 import io.camunda.client.api.command.AgentInstanceHistoryContent.ObjectContent;
 import io.camunda.client.api.command.AgentInstanceHistoryContent.TextContent;
+import io.camunda.client.api.command.AgentInstanceHistoryItem;
 import io.camunda.client.api.command.AgentInstanceHistoryMetrics;
 import io.camunda.client.api.command.AgentInstanceHistoryToolCall;
 import io.camunda.client.api.command.AgentInstanceLimits;
@@ -27,6 +28,7 @@ import io.camunda.client.api.response.DocumentMetadata;
 import io.camunda.client.api.response.DocumentReferenceResponse;
 import io.camunda.client.protocol.rest.AgentInstanceDocumentContent;
 import io.camunda.client.protocol.rest.AgentInstanceHistoryItemMetrics;
+import io.camunda.client.protocol.rest.AgentInstanceHistoryRoleEnum;
 import io.camunda.client.protocol.rest.AgentInstanceMessageContent;
 import io.camunda.client.protocol.rest.AgentInstanceObjectContent;
 import io.camunda.client.protocol.rest.AgentInstanceTextContent;
@@ -214,5 +216,38 @@ final class AgentInstanceHistoryMapper {
         .maxTokens(limits.getMaxTokens())
         .maxModelCalls(limits.getMaxModelCalls())
         .maxToolCalls(limits.getMaxToolCalls());
+  }
+
+  static io.camunda.client.protocol.rest.AgentInstanceHistoryItem toProtocolHistoryItem(
+      final AgentInstanceHistoryItem item) {
+    ArgumentUtil.ensureNotNull("historyItemId", item.getHistoryItemId());
+    if (item.getHistoryItemId().trim().isEmpty()) {
+      throw new IllegalArgumentException("historyItemId must not be blank");
+    }
+    ArgumentUtil.ensureGreaterThan("loopIteration", item.getLoopIteration(), 0);
+    ArgumentUtil.ensureNotNull("role", item.getRole());
+    ArgumentUtil.ensureNotNull("content", item.getContent());
+    ArgumentUtil.ensureNotNull("producedAt", item.getProducedAt());
+
+    final AgentInstanceHistoryRoleEnum protoRole =
+        AgentInstanceHistoryRoleEnum.fromValue(item.getRole().name());
+    if (protoRole == null) {
+      throw new IllegalArgumentException("Invalid role: " + item.getRole());
+    }
+
+    return new io.camunda.client.protocol.rest.AgentInstanceHistoryItem()
+        .historyItemId(item.getHistoryItemId())
+        .loopIteration(item.getLoopIteration())
+        .role(protoRole)
+        .content(toProtocolContent(item.getContent()))
+        .toolCalls(toProtocolToolCalls(item.getToolCalls()))
+        .metrics(toProtocolMetrics(item.getMetrics()))
+        .producedAt(item.getProducedAt().toString())
+        .tools(toProtocolTools(item.getTools()))
+        .model(item.getModel())
+        .provider(item.getProvider())
+        .limits(toProtocolLimits(item.getLimits()))
+        .systemPrompt(
+            item.getSystemPrompt() != null ? toProtocolContent(item.getSystemPrompt()) : null);
   }
 }
